build.sh: add condition if $clang_cache is not empty

+++ build.sh
@@ -175,7 +175,9 @@ if [[ ! -d "$CLANG_PATH/bin" || ! -f "$CLANG_PATH/VERSION" || "$(cat "$CLANG_PAT
 
     echo "$CLANG_INFO" > "$CLANG_PATH/VERSION"
     clang_cache=$(gh cache list | grep 'clang' | awk '{print $1}')
-    gh cache delete $clang_cache
+    if [[ -n $clang_cache ]]; then
+        gh cache delete $clang_cache
+    fi
 else
     echo "✅ Using cached Clang: $CLANG_INFO."
 fi
